+++ frontend/src/__tests__/components/CopyURL.test.js
@@ -5,6 +5,7 @@ import CopyURL from "../../components/CopyURL";
 describe("<CopyURL/>", () => {
   it("renders properly", () => {
     const { asFragment } = render(<CopyURL />);
+
     expect(asFragment()).toMatchSnapshot();
   });
 });
