+++ frontend/src/__tests__/components/DisplayPost.test.js
@@ -1,27 +1,32 @@
 import React from "react";
 import { Router } from "react-router-dom";
 import { createMemoryHistory } from "history";
-import { render } from "@testing-library/react";
+import { render, cleanup } from "@testing-library/react";
 import DisplayPost from "../../components/DisplayPost";
 
 describe("<DisplayPost/>", () => {
+  let component;
   const history = createMemoryHistory();
 
-  const component = (
-    <Router history={history}>
-      <DisplayPost />
-    </Router>
-  );
+  beforeEach(() => {
+    component = render(
+      <Router history={history}>
+        <DisplayPost />
+      </Router>
+    );
+  });
+
+  afterEach(cleanup);
 
   it("renders correctly", () => {
-    const { asFragment } = render(component);
+    const { asFragment } = component;
 
     expect(asFragment()).toMatchSnapshot();
   });
 
   it("initializes with `loading` state", () => {
-    const { getByText } = render(component);
-    const loadingText = getByText(/loading/i);
+    const { getByTestId } = component;
+    const loadingText = getByTestId("loading");
 
     expect(loadingText).toBeInTheDocument();
   });
